Done: Hoàn tất cơ bản, chạy được và ổn.

+++ Models/DataManager/UserManager.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using MitechCenter.vn.Models.Repository;
 
@@ -14,19 +14,41 @@ namespace MitechCenter.vn.Models.DataManager
 
         public void Add(User entity)
         {
+            entity.username = entity.username.ToLower();
+            entity.encryptPassword = Password.Encrypt(entity.encryptPassword);
             _context.Users.Add(entity);
             _context.SaveChanges();
         }
 
         public void Delete(User entity)
         {
-            _context.Users.Remove(entity);
-            _context.SaveChanges();
+            if (entity.username == "admin")
+            {
+                throw new System.Exception("Không thể xóa quản trị viên mặc định");
+            }
+            else
+            {
+                _context.Users.Remove(entity);
+                _context.SaveChanges();
+            }
         }
 
         public User Get(long id)
         {
             return _context.Users.FirstOrDefault(e => e.uId == id);
+        }
+
+        public User Find(string username, string password)
+        {
+            var found = _context.Users.First(u => u.username == username.ToLower());
+            if (found.username == "admin" && found.encryptPassword == "admin")
+            {
+                return found;
+            }
+            else if (found.encryptPassword == Password.Encrypt(password))
+                return found;
+            else
+                return null;
         }
 
         public User Get(string key)
